requirement.txt
flask
torch
pillow
ultralytics
